Pull the storage part of SecretStorage out to a new class
